% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering_reports.R
\name{body_add_blank_lines}
\alias{body_add_blank_lines}
\title{Add blank lines in the MS Word document}
\usage{
body_add_blank_lines(doc, n = 1)
}
\arguments{
\item{doc}{'Word' document object}

\item{n}{Integer. Number of blank lines to add.}
}
\value{
Updated 'Word' document object
}
\description{
Add blank lines in the MS Word document
}
