% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pflow.R
\name{code_target}
\alias{code_target}
\title{Generate a R code file path}
\usage{
code_target(file_name)
}
\arguments{
\item{file_name}{of the R code file.}
}
\value{
R code file text in console.
}
\description{
Generate a R code file path
}
