% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{contains_quarto}
\alias{contains_quarto}
\title{Check if quarto library call exists in packages.R}
\usage{
contains_quarto(filepath)
}
\arguments{
\item{filepath}{of packages.R}
}
\value{
a boolean
}
\description{
Check if quarto library call exists in packages.R
}
