% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering_reports.R
\name{create_title_page}
\alias{create_title_page}
\title{Create a title page of a report as a 'Word' object}
\usage{
create_title_page(
  project_name,
  title,
  author,
  designation = NULL,
  department = NULL,
  report_date = NULL
)
}
\arguments{
\item{project_name}{Project name}

\item{title}{Title of the report}

\item{author}{Author's name. Optionally, include your post-nominal title (MSc, PhD, etc).}

\item{designation}{Author's designation}

\item{department}{Author's department within SCRI}

\item{report_date}{Reporting date in the format "dd-mm-yyyy". If not provided, the system's date is used.}
}
\value{
A 'Word' object from \{officer\} package
}
\description{
Create a title page of a report as a 'Word' object
}
\examples{
\dontrun{
create_title_page(
  project_name = "Test Project",
  title = "Test Title",
  author = "Test Author",
  designation = "Senior Analyst",
  department = "Research Department",
  report_date = "01-01-2024"
)
}
}
