% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pflow.R
\name{qmd_target}
\alias{qmd_target}
\title{Generate a QMD file path}
\usage{
qmd_target(file_name)
}
\arguments{
\item{file_name}{of the QMD file}
}
\value{
a QMD file path in console
}
\description{
Generate a QMD file path
}
