% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering_reports.R
\name{render_report}
\alias{render_report}
\title{Render the input report RMarkdown file to MS Word document}
\usage{
render_report(rmd_file, add_title_page = FALSE)
}
\arguments{
\item{rmd_file}{File path of a RMarkdown file}

\item{add_title_page}{If FALSE (default), no title page added. To add one, specify details of a title as a list.}
}
\value{
File path of the resulting MS Word document.
}
\description{
Render the input report RMarkdown file to MS Word document
}
\details{
In the list for the title page details, At least `project_name`, `title`, and `author` must be specificed. `designation` and `department` are optional. `report_date` must be in the format "dd-mm-yyyy"; if not provided, system's date is used.
}
\examples{
\dontrun{
render_report(
  rmd_file = "example.Rmd",
  add_title_page = list(project_name = "Test Project",
                        title = "Test Title",
                        author = "Test Author",
                        designation = "Senior Analyst",
                        department = "Research Department",
                        report_date = "01-01-2024")
)
}
}
