% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pflow.R
\name{use_code}
\alias{use_code}
\title{Create a code file and generate a R code file path}
\usage{
use_code(file_name)
}
\arguments{
\item{file_name}{of the R code file.}
}
\value{
the path of the file created.
}
\description{
Create a code file and generate a R code file path
}
