% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pflow.R
\name{use_html}
\alias{use_html}
\title{Create a QMD file for a HTML file and generate a file path}
\usage{
use_html(file_name)
}
\arguments{
\item{file_name}{of the QMD file}
}
\value{
the path of the file created (invisibly).
}
\description{
Create a QMD file pre-configured to generate a HTML file.
}
