% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pflow.R
\name{qmd_target}
\alias{qmd_target}
\title{Generate a QMD file path}
\usage{
qmd_target(file_name, output_format_type)
}
\arguments{
\item{file_name}{of the QMD file}

\item{output_format_type}{of the QMD file; either 'docx' or 'html'}
}
\value{
a QMD file path in console
}
\description{
Generate a QMD file path
}
